Remove dead code for segment request validation
This segment request validation appears to never have been activated.

This will be solved differently in the future, by having a separate
segment request handler for local requests (TCP only) and, in core ASes
only, a simpler handler for requests from other path servers (over
SCION/QUIC).
This will simplify the validation of the requests, so the existing code
would not help a lot.