Add adapters for converting slayers data to app data
Add decoders and serializers going from SCION v2 header
raw data to spkt.ScnPkt and spath.Path.The APIs
are identical to the SCION v1 header raw data decoders and
serializers, making switching to them trivial.

Also adds the header_v2 feature flag, which should be
used by application that want to switch in code between SCION
header V1 and V2.

Performance is not optimized, and the new adapters
are expected to put a fair bit of pressure on the GC.

Also:
- Make decoding benchmarks in slayers stop on errors; this
fixes an issue where decoding was silently stopping very early
(after the SCION header) due to a missing CanDecode property;
the performance impact of the additional parsing is negligible
- slayers SCION/UDP layer now parses the SCION/UDP gopacket layer
instead of UDP
- slayers documentation includes payload extraction