log: remove Trace,Warn,Crit levels
This is a BREAKING CHANGE, because logging infrastructure might need to
re-evaluate scripts/alerting/logging levels in configs, etc. also this commit
removes the logdog tool.

All Trace logs are converted to Debug.
All Crit logs are converted to Error.
All Warn logs are converted to Info.

Having more than those remaining 3 levels is not really needed:
- debug: Contains low level details for developers.
- info: Contains general information about the process, failures that
  happen as part of normal operation are also logged at this level.
- error: Contains unexpected errors that lead to mal-functioning of the
  system in severe cases the application may choose to exit after an
  error log.

This PR also removes logdog since that was a tool that was seldomly used
and is strictly tied to the log format we have now, the format might
change in the future.

The new recommended logging level for production is info.

Further audits are needed of what we need to log at info level instead
of debug/error.

Fixes #3595