spki: make cert renewal easier to use (#3798)
Improve the usability of the cert renewal tool.

No longer require the `out` and `template` flags.
No longer require the positional argument that indicates the CA to send
the certificate chain request to.

If the `template` flag is not set, the subject of the CSR is filled with
with the subject of the transport certificate chain.

If the `out` flag is not set, the resulting chain is written to the
directory of the transport certificate chain as `ISDx-ASy.s.pem`,
where `x` is the ISD, `y` is the AS number, and `s` is the hex encoded
serial number of the AS certificate in the certificate chain.

If the positional argument is not set, the CSR is sent to the subject
of the CA certificate in the transport certificate chain.

In case verification fails against the TRC, the chain is now still
written to disk with the suffix `.unverified`. This ensures that the
CS will not automatically pick up the chain. However, operators can
still inspect the resulting chain, and check if the verification error
is the result of using the wrong TRC.

Improve the help message to describe the behavior of the tool
and to describe the expected input for the template file.