combinator: enable header v2 path combination
Add combinator logic that is capable of handling v2 segments.

The combinator decides whether a segment is v1 or v2 based on the
value in the ISD field of the control plane segment. In v1, this
field is always non-zero, in header v2 this is always zero.

The SCION Daemon API is extended such that it attaches the header version
to the replies. Client logic then decides whether to initialize the
path as header v1 or header v2 path.