cs: prepare beaconing for header changes (#3806)
Prepare the beaconing package to be able to deal with using the two
path header formats.

The beacon extender is abstracted into an interface. When support
for header v2 is added, go/pkg/cs just needs to inject the correct
beacon extender.

The onehop.Sender is abstracted into an interface. When support
for header v2 is added, go/pkg/cs just needs to inject a one hop
sender that conforms to the new format.

The SegmentRegistration RPC is abstracted into an interface that has
no resemblance with infra.Messenger. A compat adapter is introduced
in beaconing/compat. Thus, a messenger that uses the new header format
can be injected.