seg: support header v2 info and hop field (#3808)
Add timestamp and segment ID to the `PathSegmentSignedData` to support
header v2 development.

Parse the raw legacy info field when parsing the path segment
and populate the timestamp.

This allows us not to return an error in the `Timestamp` method
and makes the API of seg.PathSegment a lot easier to use.

Remove `PathSegment.InfoF` as it taints everything with header v1
details. In most places it is only used for the timestamp anyway.

The control server now beacons all beacons with the segment info for
header v1 and header v2. The remaining part is to create the correct
hop fields.

Decouple the data plane representation from the control plane
representation of the hop field.

Introduce the `hopField` entry in the capnp segment representation
that will be used to store the new hop field representation.

For legacy hop fields, the new HopField struct is used to store the
parsed values. Thus, after parsing the path segment, even if it is
for a legacy path, the raw hop field no longer has to be accessed.
This comes with the great benefit, that `ID` and `FullID` computation
can no longer return an error. Further, because of the decoupling,
completely switching to the new header v2 will be a lot easier.