ci: shorten step names and merge steps (#3824)
In buildkite, when you try to track down by name the acceptance test that you are working on, it is very hard to find it because the long names are truncated in the buildkite view. This commit shortens the long common prefix to resolve the above issue.

The convention is
 UT: Unit Tests
 IT: Integration Tests
 AT: Acceptance Tests

In this commit also, a number of steps were merged to reduce the number
of agents required per run and therefore reduce the overall running time per CI run.