cs: purge global state
Purge the global state from the control server.
Drop the hook system work around.

As part of this pull request, the control server is moved to cobra.
The flags slightly changed compared to the previous implementation.
With the move to cobra, the flags are POSIX style, and no longer go style.

This means `-config` is no longer a valid flag. Instead, use:

    cs --config cs.toml

The sample file generation has been moved to a sub command:

    cs sample config
    cs sample policy

Auto completion support is added as a sub command:

    cs completion --help