cmd: reduce duplication
Share the common commands across different binaries

The following commands are shared where appropriate.

- <bin> completion
- <bin> version
- <bin> sample
- <bin> sample config

Further, commands are refactored to be easier to compose by taking
a pather that provides the path for example in the help message.

With this change, the sciond CLI is slightly different:

`sciond --help_config` is now `sciond sample config`