disp: support SCMP informational messages
Add support for SCMP echo and traceroute messages to the dispatcher.

When the dispatcher receives a SCMP informational message request from
the application, it registers the identifer for mapping back the reply,
and sends the message to the next hop.

When the dispatcher receives a SCMP information message request from
the network, it reverses the packet and sends it back over the network.
During reversal, the requests are translated to replies.

When the dispatcher receives a SCMP information message reply from
the network, it looks up the application to deliver based on the
identifier.

Currently, the dispatcher ignores the checksum.