cs: don't verify RemoteOutIf in beaconing (#3832)
Also for v2 don't set it.

Since we want to get rid of keepalives, which also detect remote
interfaces we have to stop using remote interface IDs. Except for
peering links we don't really need to know the remote interface ID. For
peering links we should add the remote ID to the topology file once we
implement support for peering links.