disp: support SCMP error messages
Add support for delivering SCMP error messages to the application that
spawned the offending packet.

When the dispatcher receives a SCMP error message that quotes an
UDP/SCION packet, it extracts the source port of the UDP layer
and delivers the packet to the application registered with that port.

When the dispatcher receives a SCMP error message that quotes a Echo
or Traceroute request, it extracts the identifier of the quote and
delivers the packet to the application registered with that identifier.

The dispatcher does not do any sanity checks.