disp: classify decoded packet correctly
Classify the packet correctly when decoding.

If a layer does not know the type of the next layer, it will simply
return gopacket.Payload. This breaks the assumption in the dispatcher
that if a packet is parsed without error, that it must be a UDP/SCION
packet.

With this PR, the dispatcher is modified to actually check
that the decoded layer 4 is either an SCMP or a UDP. If it is non of the
two, an error is returned.