topogen: allocate ipv4 subnets for docker
If we just specify an IPv6 subnet in a docker network, docker-compose chooses
an IPv4 subnet which might overlap with another IPv4 network that we allocated.
This will make the topology fail able to start. This commit prevents this and
should fix a few CI errors.

Also remove the `--in-docker` flag and supporting code, because it is no longer needed.
Also add extra_hosts entry for jaeger tracing.