cs, sd: simplify database configs
The CS and the SD use multiple databases: path, beacon, trust,
revocation, and/or renewal. Each database was each own type
in the Go codebase, leading to a lot of duplicated code.

This PR changes the above in the following ways:
- The database configurations to reuse the same type.
- The database configuration type has been reworked from
a map to a struct.
- Helper factories have been added to a new storage package,
to simplify the creation of databases.
- CS and SD configs no longer accept database backend fields.