border: ignore EINVAL, EACCES
Analogous to scionproto/scion#3448.

When forwarding an incoming packet, `EINVAL` and `EACCESS` can be caused by
trying to send the packet to a "blackhole" or "prohibit" route.
Although not mentioned in extracts from the manual below, `EINVAL` is also
returned for any "martian" addresses during routing; specifically this occurs
when trying to send on a non-loopback device with a localhost source address
(by default, anyway).

As the destination address is controlled by an end host outside the AS,
the router should not abort in this case.

----

Relevant snippets from man **ip**(7):

    EINVAL
      Invalid argument passed.
      For send operations this can be caused by sending to a blackhole route.

    EACCES
      The user tried to execute an operation without the necessary permissions.
      These include: sending a packet to a broadcast address without having the
      SO_BROADCAST  flag  set; sending a packet via a prohibit route;
      [...]

C.f also "Route types" under **ip-route**(8):

    Route types:

      [...]

      unreachable - these destinations are unreachable. Packets are discarded
      and the ICMP message host unreachable is generated.  The local senders get an
      EHOSTUNREACH error.

      blackhole - these destinations are unreachable. Packets are discarded
      silently.  The local senders get an EINVAL error.

      prohibit - these destinations are unreachable. Packets are discarded and
      the ICMP message communication administratively prohibited is generated. The
      local senders get an EACCES error.

      [...]