snet: move MTU/Expiry to separate snet.PathMetadata interface (#3839)
Add additional interface for (optional) path metadata on snet.Path.
This new snet.PathMetadata interface will be extended to add support for
the information in the StaticInfo beacon extension (Geo information,
Latency, Bandwidth, etc) as a next step.

Having a separate interface for the metadata clarifies that some
instances of snet.Path (e.g. the `partialPath` constructed from raw
forwarding information) do not have any of this metadata available.
Thus, moving the existing MTU and Expiry fields to this new interface
feels quite natural.

Additionally, the path fingerprint can be implemented in terms of
snet.Path, so it does not need to be part of the interface.