proto: define SignedMessage
Define signed message in protobuf.
This will be used for signing AS entries and certificate renewal requests.

In addition to the protobuf definition, `go/lib/scrypto/signed` is introduced.
The library provides a safe API to sign and verify signed messages.
Future use of signed messages will always go through this library.