proto: translate path segment
Translate path segment to protobuf format.

A lot of duplicate/redundant/obsolete information is dropped with the
move from capnp to protobuf.

The only extension that is currently planed to be ported is
static info, but this will be done at a later stage, since this
can be done in a backwards compatible change.