seg: switch path segment representation from capnp to protobuf (#3874)
Switch the path segment representation from capnp to protobuf.

The protobuf representation makes a clean split between control plane and data
plane representation of a hop field. We do no longer encode the data plane wire
format in the control plane path segments.

`go/lib/ctrl/seg` is now only the application representation of the path
segments. The application representation can be encoded in the wire format using
`seg.PathSegmentToPB`. The applications are responsible for serialization and
deserialization themselves. This is no longer handled be the `seg` package. The
functionality is moved to the `pathdb` and `beacon` packages.

Signatures are now created more cleanly with the concept of associated data.

Not all extensions are ported. Some are destined to disappear, and some are not
yet ported. For example the static info extension will be ported at a later
stage. Thus, a lot of code regarding static info was disabled.

Even after this PR, code is still tainted by path_mgmt in a lot of places, but
it is now mostly used as a container for segments and revocations. When we drop
header v1 support, and we get rid of revocations, we can start removing
path_mgmt everywhere