router: SCMP parameter point to start of block
For most errors we already pointed to the beginning of the block, but
for bad packet length we pointed to the actual byte offset. The problem
with the byte offset is that not every field is by aligned. The error
code usually give enough information to determine where the problem is.