docker: remove su-exec (#3883)
Use user for running the images instead.

Using su-exec complicated the usage of our images.
Also it is not really clear what the advantages over using docker --user
is. A downside of using `docker --user` is that file permissions have to
be set before starting the image.

Also remove the root checking code in the binaries. The check was flawed
and required to bind mount an actual user in the docker containers.