StaticInfoExtension to grpc,  new metadata combinator
Add all the previously described metadata, latency, link type,
geographical information, bandwidth, number of AS internal hops and
generic notes.  Simplify & streamline the representation of the data by
using maps instead of lists, in both the protocol and the related
higher-level go structs.

Re-implement the extraction of the metadata in the path combinator.
Instead doing the complicated walk over the pathSolution with explicitly
handling cross over and peering etc, use the already created path to
guide this. Converts the data directly into a form that will eventually
be the representation used in snet.PathMetadata. Now explicitly handles
conflicting information.