unify path types in combinator, segutil, sciond (#3893)
* Combine mostly equivalent implementations of `snet.Path`
segutil.path and sciond.Path, in a shared lib in `go/lib/snet/path`.

* Adapt combinator to return a Path that is already almost identical
with this path -- the only missing bit is the UnderlayNextHop (which is
outside of the scope of the combinator, not having access to the local
topology).
In particular, the combinator now directly builds a raw forwarding path
instead of returning a high level description of the same -- using the
structs of the slayers for this actually allows to get rid of most of the
intermediate helper types.
This makes the places that call the combinator fairly trivial.

* Side quest: strip down the public interface of the combinator to the
essentials (combinator.Combine and combinator.Path).

* Drop sciond.FwdPathMeta etc, use snet.Path directly
in the sciond grpc.Daemon server. This is nice because the "input" on the
server side and the output on the client side are now the same type.

* Drop various unused sciond helper types

* Remove sciond staticinfo "condenser" and seperate types. Instead of
this, the staticinfo metadata will be added to snet.PathMetadata.
This data will be prepared and returned in the identical format directly
by the path combinator -- this is the motivation for this refactoring!

* Remove unused sciond/fetcher/filter (instead of fixing the tests)

Planned next steps (future PRs):
* add the "staticinfo" information to the `snet.PathMetadata`, adapt the existing (currently unused) code in the combinator to fill this information, adapt the sciond grpc proto to carry this extended metadata.
* change `snet.PathMetadata` from an interface to a struct. Similar to the `PathInterface`, there simply seems no reason for this to be an interface.
Depending on how exactly the extended "staticinfo" data is represented, this might be useful to allow defining higher-level accessor functions to aggregate the information but still giving applications the full information (e.g. storing latency for each hop, but allowing a simple `TotalLatency()` getter) -- note that having to deal with missing information may require the applications to have more insight than only the pre-condensed data.   
Perhaps move the `Interfaces` list into this Metadata struct as well, as it seems to belong in here.
* perhaps, attempt to turn `snet.Path` into a struct as well -- it seems that the behaviour of all the different implementations (empty path, partial path, ...) can easily be combined.
* remove the now pointless `segutil.Router`

Closes #3872