posix-gateway: config file rework
Various changes to configuration file are introduced:

The posix-gateway specific tunneling configuration is extracted in a separate
`[tunnel]` section. The behavior of `ctrl_addr` and `data_addr` is now more in
sync with other definitions of `addr`, i.e., host:port, instead of just IP. See
the config help for more information on `ctrl_addr` and `data_addr`.

Renames:

- `[sig]` is renamed to `[gateway]`
- `sig.sig_config` is renamed to `gateway.traffic_policy_file`
- `sig.tun` is renamed to `tunnel.name`
- `sig.tun_routing_table_id` is renamed to `tunnel.routing_table_id`