Move Interfaces to snet.PathMetadata
Requires changes to pathpol.PathSet -- now the pathpol is fed directly
with the Metadata struct. This seems quite natural, as it's conceivable
that the path policy would access the various metadata items in the
future.
This change also requires some minor adjustments at the call sites for
the path policies.