cmds: usage string based on binary name
Renaming binaries is sometimes preferrable or even necessary when
creating distribution packages. In order to reduce confusion for the
users, it's nice when the built in help text of the applications refers
to the actually used name of the binary.

By adapting the `Use` string, cobra also generates correct completion
scripts when using a changed name for the binary.