Added SRV only DNS request
Currently the DNS-SD mode is wrongly used, with respect to how it is
described it should work in the thesis and the online documentation.
In fact, while the documentation suggests to use a SRV only DNS backed
hinting, the code suggests the usage should be of canonical DNS-SD.
We keep the current code but also add a SRV only request that, IMHO, is
all we need for the current use case and also for future use cases.