scion-pki: fix combine TRC flake
With the new go version, the SignerInfos field is finally sorted according to X690 Section 11.6.

scion-pki already sorts the SignerInfos to produced deterministic output. Now, the order of your custom sorting code is different then the one defined by X690 Section 11.6. This leads to a flake in the unit tests.

We leave the custom sort in the code, such that scion-pki compiled with an older go version will still output deterministic TRC SignerInfos.

In go1.15 and later, the SignerInfos field is sorted during final serialization, and the custom order will be replaced by the one defined in X690 Section 11.6.

This does not pose a problem for signature verification or creation and should not affect TRCs created with previous go versions.