hp: Simplify configuration
This changes the hidden paths design doc to simplify configuration
and operations. The multiple files are coalesced in a single
hidden paths configuration that can be loaded by both Daemons and
Control Services. For Control Services that need additional policies,
the file includes an optional registration configuration section.

This should also simplify code, because configuration can now be made
atomic. The previous design needed a lot of data correlation logic
between information coming from groups files and the registration
policy.

Some fine-tuning parameters that are not needed for the first
version of hidden paths (like expiration and up-segment registration)
have been removed in the interest of keeping it simple.

Also, the JSON format is changed to YAML so that all configuration
is now consistent.

[doc]