cs: Decrease timeouts in beacon origination/propagation
Fixes a bug where the periodic task has a very long ctx deadline (equal
to the origination periodicity). While the topology is unstable at the
start, origination RPCs time out. However, because the ctx is extremely
long-lived, they time out after the period (i.e., after 5 minutes). So
the 500ms loop timer does nothing, nothing is rechecked, and the main
worker is stuck waiting for its work group to be done

This makes it so that RPCs time out after 10 seconds of silent time,
allowing the beaconing and propagation tasks to retry sooner.

Also changes beaconing.Tick to no longer mix method writers with direct
external writes by creating an accessor API. This makes it evident that
if there are internal invariants in the type's methods, they are enforced
by the implementation.