gateway: simplify defaults for policy file config (#3941)
The interpretation of the configured `ip_routing_policy_file`
configuration option was:
- if the value is identical to the default file path and the file does
  not exist, ignore it and return a default routing policy
- otherwise, read the file at the specified location (and fail if it
  doesn't exist)

Not only does this seem a bit convoluted, but it can also fail in
interesting ways; when the user does not want to specify an
`ip_routing_policy_file` and there is an error stat-ing the file at the
default location (e.g. no permission to read directory, error while
accessing network file system, etc.), the SIG would refuse to start.

Simplified to only handle the empty value as special default.
The consequence of this is that the configuration files now _must_
specify the `ip_routing_policy` file path if it should be loaded.

Also, removed the default for the `traffic_policy_file` option, making
this option mandatory in the configuration file.

Finally, removed the unused `dispatcher` configuration entry.