pathprobe: move go/lib/sciond/pathprobe to go/pkg/pathprobe
The package pathprobe is moved from go/lib/sciond/pathprobe  to go/pkg/pathprobe.

Package pathprobe is a general-purpose package, and not tied to the daemon. The daemon does not use this package at all (and it would be odd for it to use it, because the package is about SCION traffic and the daemon does not speak over SCION).