doc: update beacon metadata extension docs
Update beacon metadata docs after various changes to the implementation,
in particular changes done to the data representation when converting the
beacon extension to protobuf/gRPC.

Rewrite parts of the introduction, add (better) illustrations for the
data included in the beacons.
The new figures were created with draw.io, the source file is
`beacon-metadata.xml`.