lint: bazelify go linting
This is a first draft of how bazelified linting of code can look like.
This has the advantage that by default all go_library and go_test code
is checked and we don't have to remember to adapt the call in the
scion.sh invocation when we add a new directory.

GitOrigin-RevId: 110fe6cfde3d78867bca60d231f98ecde0551def