beacon metadata: fix core x-over, validate config
Fix the intra-AS metadata included for cross-over from up-segments
to core-segments, to match the documented scheme.
Concretely, the core segments now contain no data for cross-over.
Instead, the up/down segments now contain cross-over information to
_all_ core interfaces (previously only a subset).
Previously, core-segments were handled just like down segments, without
taking into account that cross-over to the _ingress_ interface of the
core-segment beacon also needs to be ensured; consequently, the data for
the cross-over between the up- and core-segment was missing.

Also, add some validation/convenience functionality for the
configuration read from the staticInfoConfig.json file. Specifically,
this extends missing information from the "Intra" sections to be
symmetric (symmetry is an assumption of the system).