posix-gateway: fix go routine leak in dataplane session
Fix a go routine leak in the dataplane session.

Before this patch, the gateway would leak go routines whenever paths
are updated, or when path updates fail.

With this patch, both of these cases are handled, and no go routines
are leaked in `Session.SetPaths` anymore.

Additionally, this patch fixes a bug where tiny MTU would cause a
uint16 underflow, resulting in a huge frame size.

GitOrigin-RevId: 30e7715fe84820f0c39384497c85cbcc9509f90b