storage: move trust DB implementation
Move the trust database implementations under `pkg/storage` because they
conceptually do not belong in the `trust` package.

The `trust` package defines a narrow interface for what it expects from a `DB`.
`storage.TrustDB` on the other hand is a super set of it, and is implemented by
the moved packages.

This PR is done in preparation for extending the trust database with
observability functionality that is outside of what `trust.DB` defines.

GitOrigin-RevId: d11f1970d8ae99edb0076558b0440d9a3519bd90