cppki: use appropriate message digest size
This change is a preparation step to follow the common recommendation
to use the hash algorithm that produces the message digest of the same
bit size as the public key algorithm for ECDSA.

We extend the set of acceptable signature algorithms to:
- ecdsa-with-SHA256
- ecdsa-with-SHA384
- ecdsa-with-SHA512

The control plane message signer is adapted to use the appropriate
signature algorithm when signing path segments and other messages.

This change also reverts some modifications to the vendored CMS code.
In particular, the hash algorithm is now chosen one-to-one the same as
in the upstream.

The changes that switches the behavior in the CA, and updates the documentation
is done in a follow-up story. This allows a staggered roll-out if necessary.

sources:
The [BSI recommendations (section 4.2)](https://www.bsi.bund.de/SharedDocs/Downloads/EN/BSI/Publications/TechGuidelines/TR03111/BSI-TR-03111_V-2-0_pdf.pdf?__blob=publicationFile&v=2) are to use a hash size of the same length as the key size.
The [Mozilla root store policy (section 5.1.2)](https://www.mozilla.org/en-US/about/governance/policies/security-group/certs/policy/) even forbids ECDSAWithSHA256 for a key of length 384.
GitOrigin-RevId: 339e4466da5615ad8d147e16d5d1e3926c92ac9b