cs: more robust retry mechanism for seg requests
The current path management mechanism for segment requests is random
selection and retry. Previously, the full request deadline was divided
into three equal time slots and (at most) three attempts were made.
This has two problems:
- Each time slot is short and potentially simply too short to make the
  entire request (in particular for high latency connections). Thus,
  every try may fail even if it hit a working path and if the full
  request deadline would leave enough time to succeed.
- We give up after three failed attempts and do not use the
  remaining time until the request deadline to probe more paths.

These two main changes attempt to alleviate these problems and increase the
chances of success:
- Limit only the time for dialing a gRPC connection for each try; once
  it has dialed succesfully, we give the entire request deadline to each
  try.
- Allow an unlimited number of retries within the request deadline.

Additionally, the Dialer now uses `OpenStream` instead of
`OpenStreamSync`. This avoids an unnecessary roundtrip during dialing
which also contributes to increasing the chances of a successful request
within the request deadline.

Note: none of this is really relevant for segment requests from sciond
to CS as these go locally via TCP. This entire retry mechanism is in the
shared logic though and so it is still applied there. This should be
harmless.