router: optimize metrics use during forwarding
Optimization based on profiling results, very low hanging fruit.

Instantiating the metrics with labels is quite costly. As this was done
for each packet during packet forwarding, this accounted for a
significant fraction of the time spent per packet.
The exact numbers probably depend on the setup but in my case I saw
about 20% of total time spent in creating the labels (with a significant
fraction of the time spent in `IA.String()`) and instantiating the
counters (i.e. `With(labels)`). To put this in perspective, this about
twice as much as the time spent in the main `processPkt`.