cs: set the correct TRC ID in the Verification Key ID
With this change, the control service sets the `trc_base` and `trc_serial`
in the `VerificationKeyID` to the latest TRC that is available, in accordance with
the [documentation](https://internal.docs.anapaya.net/en/latest/src/cryptography/interactions.html#signing-process).
Previously, it was set to the latest TRC that authenticates the signer's certificate chain.

This change is necessary to ensure that a TRC update is propagated in the network, even if the control service
does not yet have a certificate chain that is verifiable against the new TRC.

This is the correct behavior that is also required by the proto definition:

```
// VerificationKeyID is used to identify certificates that authenticate the
// verification key used to verify signatures.
message VerificationKeyID {
    // ISD-AS of the subject.
    uint64 isd_as = 1;
    // SubjectKeyID referenced in the certificate.
    bytes subject_key_id = 2;
    // Base number of the latest TRC available to the signer at the time of
    // signature creation.
    uint64 trc_base = 3;
    // Serial number of the latest TRC available to the signer at the time of
    // signature creation.
    uint64 trc_serial = 4;
}
```

GitOrigin-RevId: 3e0af7efa72f3ea64e7b0fadf182c4dce225628e