router: async read/write
Decouple reading and processing input packets from writing messages back
out. This removes the blocking write from the reader loop. Output
messages are transferred from the reader to the writer asynchronously,
via a buffered channel.
If the writer is the slowest part, packets will now be dropped
internally -- previously, the blocking write would have caused the
router to stop reading and the packets to be eventually dropped by the
OS as it ran out of UDP rmem.