trust: update committed test crypto material
The test crypto material we committed over a year ago finally reached
its lifetime limit. This is a short term fix that recreates the test
crypto material for another year.

The proper fix is to create the necessary crypto material on the fly,
and not commit it to the source tree at all. This however requires
small small improvements to scion-pki such that we no longer need
crypto_lib.sh and the openssl docker container.

GitOrigin-RevId: 9b24a27a7f5915f157fb4549a46b33cb7a6dc846