router: simplify scmp reply handling
Simplify the SCMP reply message handling.
Previously, an explicit copy of the input packet was made, so that it
can be used in the quote for an SCMP reply. But the path header of this
updated to the current state of the packet, in order to have identical
behaviour to anapayas VPP router. Now, the input packet, in the current
state of processing, is used directly for the quote.

This somewhat puzzling step of updating the quoted packet to the current
processing state was also responsible for clearing the router alert in
traceroute replies. This is now done explicitly during processing, with
the consequence that router alerts are always cleared, as was probably
intended, also for non-traceroute packets.

Parse/skip SCION extension headers from the beginning, so that the
actual L4 protocol type is available when SCMP replies need to be sent.
Previously, the full packet was parsed again to determine whether the
input packet was an SCMP error message.
Also, handling traceroute packets now checks that the input packet
was even an SCMP message.