router: sanity checks for one-hop path
Previously used the source IA of the packet to determine whether the
packet was entering or leaving the local AS. This would seemingly have
allowed a mischievous neighbor AS to relay OHP packets to any other
external interface of a router.
Now the direction is inferred from the router interface on which the
packet was received.

Also add more specific sanity checks for src/dstIA of OHP packets.

Fix remote isd_as for interface 131 in router_multi acceptance test
topology to match br-multi.png and expectations in the code. With the
changes above, the mismatch between the config and the expectations was
causing the test to fail.