router: sanity checks for one-hop path
Previously used the source IA of the packet to determine whether the
packet was entering or leaving the local AS. This would seemingly have
allowed a mischievous neighbor AS to relay OHP packets to any other
external interface of a router.
Now the direction is inferred from the router interface on which the
packet was received.

Also add more specific sanity checks for src/dstIA of OHP packets.

Closes #4052

GitOrigin-RevId: 1cae93e71b26f640376a5bf98fc429cf538398e8