slayers: fix padding after TLV options
Padding after the last TLV option is responsible for bringing the total length of the extension header to a multiple of the `LineLen`, i.e. to a multiple of 4 bytes.
The calculation for this padding was slightly wrong. It added `length % LineLen` instead of `LineLen - length % LineLen` padding bytes.
As a consequence, the total size of the extension was not padded to a multiple of 4 when a TLV option with odd length was used, triggering the corresponding error later during serialization.

Closes #4065

GitOrigin-RevId: 1202cba7cd02be6bbae1c418b6f0dcea54650d5e