slayers: remove gopacket/layers dependency
The `slayers` library, which is used by `snet` and therefore by most SCION applications, was importing the `gopacket/layers` library. This library contains a considerable amount of bloat as it supports parsing every protocol under the sun. This was used only for a small amount of basic functionality that can easily be replicated (like `BaseLayer` and the some parts of the SCION/UDP layer), as well as to support SCION/BFD. The BFD support is not needed by applications other than the router.
Fortunately, `gopacket`'s layer type system allows to define the BFD layer type as a constant, so that the `slayers.SCION` `NextLayerType` can support BFD without actually linking against `gopacket/layers`. (Note: this `NextLayerType` is not even used in the router as it handles BFD packets separately. It is however used in various test cases that parse BFD packets with `NewPacket`).

This change removes the `gopacket/layers` dependency from `slayers`. For applications that import `gopacket/layers` explicitly, e.g. because they need to parse SCION/BFD packets, the functionality remains unchanged. For applications that do not import `gopacket/layers`, SCION/BFD packets can no longer be parsed, and the binary size is reduced by by about 2MB.

Closes #4074

GitOrigin-RevId: 5587a2b67be13a3b9f59a54566d720e15fb44236