refactor: remove SignedRevInfo and capnproto dependency
The signed revocation mechanism has previously been removed from the SCION protocol and replaced by simpler (unsigned) Interface Down messages. Internally, the `SignedRevInfo` objects were still created to interface with various components that had not yet been updated during this change, in particular for the internal interface of the daemon, the `RevCache` and the segment verifier.

This change removes the `SignedRevInfo` type and replaces it by directly using the corresponding `RevInfo` where appropriate. In the segverifier, the (previously unused) support for handling revocations was removed altogether.

This allows to remove capnproto as a dependency; `SignedRevInfo` was the last active use for capnproto (to create a `SignedRevInfo` object, a `RevInfo` was `Pack`ed).  Some inactive colibri code was also still relying on capnproto, but this could easily be stripped out; work to replace this with gRPC is ongoing.
Dropping the capnproto dependency reduces the size of the resulting binaries by about 2MB (for the `end2end` binary, that's a bit under 10%).

Closes #4072

GitOrigin-RevId: 76bd60c1bbb52a3bfbc640235651baaa17a40cf9